backend/requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
gdacs-api==1.0.0
geojson==3.1.0
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
scipy==1.12.0
searoute==1.3.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
xmltodict==0.13.0
yarg==0.1.9
